scripts/requirements.txt
discord==2.3.2
python-frontmatter==1.1.0
